requirements.txt
Babel==1.3
Django==1.7.3
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
Pygments==2.0rc1
Sphinx==1.3b1
Whoosh==2.6.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
asana==0.0.5
billiard==3.3.0.18
celery==3.1.16
distribute==0.6.24
django-angular==0.7.10
django-appconf==0.6
django-autoslug==1.7.2
django-celery==3.1.16
django-compressor==1.4
django-cors-headers==1.0.0
django-debugtools==1.2.1
django-extra-views==0.6.5
django-form-utils==1.0.2
django-froala-editor
django-gmapi==1.0.1
django-gunicorn==0.1.1
django-haystack==2.3.1
django-kombu==0.9.4
django-libsass==0.2
django-mail-queue==2.1.2
django-password-reset==0.7
django-reversion==1.8.5
django-suit==0.2.11
django-suit-redactor==0.0.2
djangorestframework==3.0.5
docutils==0.12
gunicorn==18.0
kombu==3.0.23
libmagic==1.0
libsass==0.6.2
psycopg2==2.5.4
python-magic==0.4.6
python-memcached==1.53
python-ptrace==0.8
pytz==2014.7
requests==2.5.1
six==1.5.2
snowballstemmer==1.2.0
sorl-thumbnail==11.12.1b
threadpool==1.2.7
wsgiref==0.1.2
xlwt==0.7.5
